LMA: Repair of HomeWorks + changing packages paths

+++ src/main/java/sloup/Z_java_basics/Calculator.java
@@ -22,7 +22,7 @@ public class Calculator {
         if (number2 == 0) {
             throw new IllegalArgumentException("Dividing by 0 is forbidden.");
         } else {
-            double result = number1 / number2;
+            int result = number1 / number2;
             log.debug("Result is: [{}]", result);
             return result;
         }
